pyproject.toml
[project]
name = "antifraud-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.122.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyyaml>=6.0.3",
    "redis>=7.1.0",
    "uvicorn>=0.40.0",
    "prometheus-client>=0.24.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
]
